moved GncParams to separate file, addressing comments by Frank, 1/n
